WIP: Porting to x86_64 architecture
8acbc2c72ed0fc861a7161db95a37c38fd8cb53e7b3410658e980dcc9817f532

- Added x86_64 architecture support.
- Project will eventually no longer support 32 bit architectures.

+++ include/drivers/mouse.h
@@ -32,8 +32,8 @@ typedef struct {
 } MouseEvent;
 
 extern MouseEvent mouse_buffer[16];
-extern volatile uint8_t buffer_head;
-extern volatile uint8_t buffer_tail;
+extern volatile uint32_t buffer_head;
+extern volatile uint32_t buffer_tail;
 
 void RARE_FUNC init_mouse();
 extern void mouse_handler();
