+++ include/drivers/storage/ata.h
@@ -72,7 +72,7 @@ extern int CTRL_LEGACY;
 
 void RARE_FUNC init_ata(pci_device_t* pci_ata_device);
 
-void ata_read_sector(uint32_t lba, uint8_t* buffer);
-void ata_write_sector(uint32_t lba, uint8_t* buffer);
+void ata_read_sector(uint64_t lba, uint8_t* buffer);
+void ata_write_sector(uint64_t lba, uint8_t* buffer);
 
 #endif
